feature(Users): refactor Enums
