+++ src/main/kotlin/com/jlima/bookstoremanager/dto/UserDTO.kt
@@ -1,9 +1,13 @@
 package com.jlima.bookstoremanager.dto
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import com.jlima.bookstoremanager.core.Gender
+import com.jlima.bookstoremanager.enums.Gender
+import com.jlima.bookstoremanager.enums.Role
 import com.jlima.bookstoremanager.helper.ValidUUID
+import org.hibernate.validator.constraints.UniqueElements
 import java.util.Date
+import javax.persistence.EnumType
+import javax.persistence.Enumerated
 import javax.validation.constraints.Email
 import javax.validation.constraints.NotEmpty
 import javax.validation.constraints.NotNull
@@ -19,6 +23,7 @@ data class UserDTO(
     @field:Size(min = 3, max = 255)
     val name: String,
 
+    @Enumerated(EnumType.STRING)
     @field:NotNull
     @field:NotEmpty
     val gender: Gender,
@@ -29,6 +34,7 @@ data class UserDTO(
 
     @field:NotNull
     @field:NotEmpty
+    @field:UniqueElements
     @field:Email
     val email: String,
 
@@ -38,7 +44,8 @@ data class UserDTO(
     @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
     val password: String,
 
+    @Enumerated(EnumType.STRING)
     @field:NotNull
     @field:NotEmpty
-    val role: String,
+    val role: Role,
 )
