+++ src/main/kotlin/com/jlima/bookstoremanager/enums/Gender.kt
@@ -1,4 +1,4 @@
-package com.jlima.bookstoremanager.core
+package com.jlima.bookstoremanager.enums
 
 enum class Gender {
     MALE, FEMALE, OTHER
