+++ src/main/kotlin/com/jlima/bookstoremanager/enums/Role.kt
@@ -1,0 +1,5 @@
+package com.jlima.bookstoremanager.enums
+
+enum class Role {
+    ADMIN, USER
+}
